+++ test/globaldb_tests.py
@@ -1,4 +1,4 @@
-#The MIT License (MIT)
+﻿#The MIT License (MIT)
 #Copyright (c) 2014 Microsoft Corporation
 
 #Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -32,18 +32,7 @@ import pydocumentdb.constants as constants
 import pydocumentdb.global_endpoint_manager as global_endpoint_manager
 import pydocumentdb.endpoint_discovery_retry_policy as endpoint_discovery_retry_policy
 
-host = '[YOUR_GLOBAL_ENDPOINT_HERE]'
-write_location_host = '[YOUR_WRITE_ENDPOINT_HERE]'
-read_location_host = '[YOUR_READ_ENDPOINT_HERE]'
-read_location2_host = '[YOUR_READ_ENDPOINT2_HERE]'
-masterKey = '[YOUR_KEY_HERE]'
 
-write_location = '[YOUR_WRITE_LOCATION_HERE]'
-read_location = '[YOUR_READ_LOCATION_HERE]'
-read_location2 = '[YOUR_READ_LOCATION2_HERE]'
-
-test_database_id = 'testdb'
-test_collection_id = 'testcoll'
 
 #IMPORTANT NOTES: 
   
@@ -54,6 +43,19 @@ test_collection_id = 'testcoll'
 #   associated with your DocumentDB account.
 
 class Test_globaldb_tests(unittest.TestCase):
+
+    host = '[YOUR_GLOBAL_ENDPOINT_HERE]'
+    write_location_host = '[YOUR_WRITE_ENDPOINT_HERE]'
+    read_location_host = '[YOUR_READ_ENDPOINT_HERE]'
+    read_location2_host = '[YOUR_READ_ENDPOINT2_HERE]'
+    masterKey = '[YOUR_KEY_HERE]'
+
+    write_location = '[YOUR_WRITE_LOCATION_HERE]'
+    read_location = '[YOUR_READ_LOCATION_HERE]'
+    read_location2 = '[YOUR_READ_LOCATION2_HERE]'
+
+    test_database_id = 'testdb'
+    test_collection_id = 'testcoll'
 
     def __AssertHTTPFailureWithStatus(self, status_code, sub_status, func, *args, **kwargs):
         """Assert HTTP failure with status.
@@ -71,23 +73,23 @@ class Test_globaldb_tests(unittest.TestCase):
             self.assertEqual(inst.sub_status, sub_status)
 
     def setUp(self):
-        self.client = document_client.DocumentClient(host, {'masterKey': masterKey})
+        self.client = document_client.DocumentClient(Test_globaldb_tests.host, {'masterKey': Test_globaldb_tests.masterKey})
         
         # Create the test database only when it's not already present
-        query_iterable = self.client.QueryDatabases('SELECT * FROM root r WHERE r.id=\'' + test_database_id + '\'')
+        query_iterable = self.client.QueryDatabases('SELECT * FROM root r WHERE r.id=\'' + Test_globaldb_tests.test_database_id + '\'')
         it = iter(query_iterable)
         
         self.test_db = next(it, None)
         if self.test_db is None:
-            self.test_db = self.client.CreateDatabase({'id' : test_database_id})
+            self.test_db = self.client.CreateDatabase({'id' : Test_globaldb_tests.test_database_id})
 
         # Create the test collection only when it's not already present
-        query_iterable = self.client.QueryCollections(self.test_db['_self'], 'SELECT * FROM root r WHERE r.id=\'' + test_collection_id + '\'')
+        query_iterable = self.client.QueryCollections(self.test_db['_self'], 'SELECT * FROM root r WHERE r.id=\'' + Test_globaldb_tests.test_collection_id + '\'')
         it = iter(query_iterable)
         
         self.test_coll = next(it, None)
         if self.test_coll is None:
-            self.test_coll = self.client.CreateCollection(self.test_db['_self'], {'id' : test_collection_id})
+            self.test_coll = self.client.CreateCollection(self.test_db['_self'], {'id' : Test_globaldb_tests.test_collection_id})
         
     def tearDown(self):
         # Delete all the documents created by the test case for clean up purposes
@@ -99,7 +101,7 @@ class Test_globaldb_tests(unittest.TestCase):
         connection_policy = documents.ConnectionPolicy()
         connection_policy.EnableEndpointDiscovery = False
 
-        client = document_client.DocumentClient(host, {'masterKey': masterKey}, connection_policy)
+        client = document_client.DocumentClient(Test_globaldb_tests.host, {'masterKey': Test_globaldb_tests.masterKey}, connection_policy)
 
         document_definition = { 'id': 'doc',
                                 'name': 'sample document',
@@ -107,7 +109,7 @@ class Test_globaldb_tests(unittest.TestCase):
         
         # When EnableEndpointDiscovery is False, WriteEndpoint is set to the endpoint passed while creating the client instance
         created_document = client.CreateDocument(self.test_coll['_self'], document_definition)
-        self.assertEqual(client.WriteEndpoint, host)
+        self.assertEqual(client.WriteEndpoint, Test_globaldb_tests.host)
         
         # Delay to get these resources replicated to read location due to Eventual consistency
         time.sleep(5)
@@ -116,20 +118,20 @@ class Test_globaldb_tests(unittest.TestCase):
         content_location = str(client.last_response_headers['content-location'])
 
         content_location_url = urlparse.urlparse(content_location)
-        host_url = urlparse.urlparse(host)
+        host_url = urlparse.urlparse(Test_globaldb_tests.host)
         
         # When EnableEndpointDiscovery is False, ReadEndpoint is set to the endpoint passed while creating the client instance
         self.assertEqual(str(content_location_url.hostname), str(host_url.hostname))
-        self.assertEqual(client.ReadEndpoint, host)
+        self.assertEqual(client.ReadEndpoint, Test_globaldb_tests.host)
         
         connection_policy.EnableEndpointDiscovery = True
         document_definition['id'] = 'doc2'
 
-        client = document_client.DocumentClient(host, {'masterKey': masterKey}, connection_policy)
+        client = document_client.DocumentClient(Test_globaldb_tests.host, {'masterKey': Test_globaldb_tests.masterKey}, connection_policy)
 
         # When EnableEndpointDiscovery is True, WriteEndpoint is set to the write endpoint
         created_document = client.CreateDocument(self.test_coll['_self'], document_definition)
-        self.assertEqual(client.WriteEndpoint, write_location_host)
+        self.assertEqual(client.WriteEndpoint, Test_globaldb_tests.write_location_host)
         
         # Delay to get these resources replicated to read location due to Eventual consistency
         time.sleep(5)
@@ -138,17 +140,17 @@ class Test_globaldb_tests(unittest.TestCase):
         content_location = str(client.last_response_headers['content-location'])
         
         content_location_url = urlparse.urlparse(content_location)
-        write_location_url = urlparse.urlparse(write_location_host)
+        write_location_url = urlparse.urlparse(Test_globaldb_tests.write_location_host)
 
         # If no preferred locations is set, we return the write endpoint as ReadEndpoint for better latency performance
         self.assertEqual(str(content_location_url.hostname), str(write_location_url.hostname))
-        self.assertEqual(client.ReadEndpoint, write_location_host)
+        self.assertEqual(client.ReadEndpoint, Test_globaldb_tests.write_location_host)
 
     def test_globaldb_endpoint_discovery(self):
         connection_policy = documents.ConnectionPolicy()
         connection_policy.EnableEndpointDiscovery = False
 
-        read_location_client = document_client.DocumentClient(read_location_host, {'masterKey': masterKey}, connection_policy)
+        read_location_client = document_client.DocumentClient(Test_globaldb_tests.read_location_host, {'masterKey': Test_globaldb_tests.masterKey}, connection_policy)
 
         document_definition = { 'id': 'doc',
                                 'name': 'sample document',
@@ -172,7 +174,7 @@ class Test_globaldb_tests(unittest.TestCase):
         }))
 
         connection_policy.EnableEndpointDiscovery = True
-        read_location_client = document_client.DocumentClient(read_location_host, {'masterKey': masterKey}, connection_policy)
+        read_location_client = document_client.DocumentClient(Test_globaldb_tests.read_location_host, {'masterKey': Test_globaldb_tests.masterKey}, connection_policy)
 
         # CreateDocument call will go to the WriteEndpoint as EnableEndpointDiscovery is set to True and client will resolve the right endpoint based on the operation
         created_document = read_location_client.CreateDocument(self.test_coll['_self'], document_definition)
@@ -182,7 +184,7 @@ class Test_globaldb_tests(unittest.TestCase):
         connection_policy = documents.ConnectionPolicy()
         connection_policy.EnableEndpointDiscovery = True
 
-        client = document_client.DocumentClient(host, {'masterKey': masterKey}, connection_policy)
+        client = document_client.DocumentClient(Test_globaldb_tests.host, {'masterKey': Test_globaldb_tests.masterKey}, connection_policy)
         
         document_definition = { 'id': 'doc',
                                 'name': 'sample document',
@@ -198,14 +200,14 @@ class Test_globaldb_tests(unittest.TestCase):
         content_location = str(client.last_response_headers['content-location'])
 
         content_location_url = urlparse.urlparse(content_location)
-        write_location_url = urlparse.urlparse(write_location_host)
+        write_location_url = urlparse.urlparse(Test_globaldb_tests.write_location_host)
 
         # If no preferred locations is set, we return the write endpoint as ReadEndpoint for better latency performance
         self.assertEqual(str(content_location_url.hostname), str(write_location_url.hostname))
-        self.assertEqual(client.ReadEndpoint, write_location_host)
+        self.assertEqual(client.ReadEndpoint, Test_globaldb_tests.write_location_host)
 
-        connection_policy.PreferredLocations = [read_location2]
-        client = document_client.DocumentClient(host, {'masterKey': masterKey}, connection_policy)
+        connection_policy.PreferredLocations = [Test_globaldb_tests.read_location2]
+        client = document_client.DocumentClient(Test_globaldb_tests.host, {'masterKey': Test_globaldb_tests.masterKey}, connection_policy)
 
         document_definition['id'] = 'doc2'
         created_document = client.CreateDocument(self.test_coll['_self'], document_definition)
@@ -217,47 +219,47 @@ class Test_globaldb_tests(unittest.TestCase):
         content_location = str(client.last_response_headers['content-location'])
 
         content_location_url = urlparse.urlparse(content_location)
-        read_location2_url = urlparse.urlparse(read_location2_host)
+        read_location2_url = urlparse.urlparse(Test_globaldb_tests.read_location2_host)
         
         # Test that the preferred location is set as ReadEndpoint instead of default write endpoint when no preference is set
         self.assertEqual(str(content_location_url.hostname), str(read_location2_url.hostname))
-        self.assertEqual(client.ReadEndpoint, read_location2_host)
+        self.assertEqual(client.ReadEndpoint, Test_globaldb_tests.read_location2_host)
 
     def test_globaldb_endpoint_assignments(self):
         connection_policy = documents.ConnectionPolicy()
         connection_policy.EnableEndpointDiscovery = False
 
-        client = document_client.DocumentClient(host, {'masterKey': masterKey}, connection_policy)
+        client = document_client.DocumentClient(Test_globaldb_tests.host, {'masterKey': Test_globaldb_tests.masterKey}, connection_policy)
 
         # When EnableEndpointDiscovery is set to False, both Read and Write Endpoints point to endpoint passed while creating the client instance
-        self.assertEqual(client._global_endpoint_manager.WriteEndpoint, host);
-        self.assertEqual(client._global_endpoint_manager.ReadEndpoint, host);
+        self.assertEqual(client._global_endpoint_manager.WriteEndpoint, Test_globaldb_tests.host);
+        self.assertEqual(client._global_endpoint_manager.ReadEndpoint, Test_globaldb_tests.host);
 
         connection_policy.EnableEndpointDiscovery = True
-        client = document_client.DocumentClient(host, {'masterKey': masterKey}, connection_policy)
+        client = document_client.DocumentClient(Test_globaldb_tests.host, {'masterKey': Test_globaldb_tests.masterKey}, connection_policy)
 
         # If no preferred locations is set, we return the write endpoint as ReadEndpoint for better latency performance, write endpoint is set as expected
-        self.assertEqual(client._global_endpoint_manager.WriteEndpoint, write_location_host);
-        self.assertEqual(client._global_endpoint_manager.ReadEndpoint, write_location_host);
+        self.assertEqual(client._global_endpoint_manager.WriteEndpoint, Test_globaldb_tests.write_location_host);
+        self.assertEqual(client._global_endpoint_manager.ReadEndpoint, Test_globaldb_tests.write_location_host);
 
-        connection_policy.PreferredLocations = [read_location2]
-        client = document_client.DocumentClient(host, {'masterKey': masterKey}, connection_policy)
+        connection_policy.PreferredLocations = [Test_globaldb_tests.read_location2]
+        client = document_client.DocumentClient(Test_globaldb_tests.host, {'masterKey': Test_globaldb_tests.masterKey}, connection_policy)
 
         # Test that the preferred location is set as ReadEndpoint instead of default write endpoint when no preference is set
-        self.assertEqual(client._global_endpoint_manager.WriteEndpoint, write_location_host);
-        self.assertEqual(client._global_endpoint_manager.ReadEndpoint, read_location2_host);
+        self.assertEqual(client._global_endpoint_manager.WriteEndpoint, Test_globaldb_tests.write_location_host);
+        self.assertEqual(client._global_endpoint_manager.ReadEndpoint, Test_globaldb_tests.read_location2_host);
 
     def test_globaldb_update_locations_cache(self):
-        client = document_client.DocumentClient(host, {'masterKey': masterKey})
+        client = document_client.DocumentClient(Test_globaldb_tests.host, {'masterKey': Test_globaldb_tests.masterKey})
 
-        writable_locations = [{'name' : write_location, 'databaseAccountEndpoint' : write_location_host}]
-        readable_locations = [{'name' : read_location, 'databaseAccountEndpoint' : read_location_host}, {'name' : read_location2, 'databaseAccountEndpoint' : read_location2_host}]
+        writable_locations = [{'name' : Test_globaldb_tests.write_location, 'databaseAccountEndpoint' : Test_globaldb_tests.write_location_host}]
+        readable_locations = [{'name' : Test_globaldb_tests.read_location, 'databaseAccountEndpoint' : Test_globaldb_tests.read_location_host}, {'name' : Test_globaldb_tests.read_location2, 'databaseAccountEndpoint' : Test_globaldb_tests.read_location2_host}]
         
         write_endpoint, read_endpoint = client._global_endpoint_manager.UpdateLocationsCache(writable_locations, readable_locations)
 
         # If no preferred locations is set, we return the write endpoint as ReadEndpoint for better latency performance, write endpoint is set as expected
-        self.assertEqual(write_endpoint, write_location_host)
-        self.assertEqual(read_endpoint, write_location_host)
+        self.assertEqual(write_endpoint, Test_globaldb_tests.write_location_host)
+        self.assertEqual(read_endpoint, Test_globaldb_tests.write_location_host)
 
         writable_locations = []
         readable_locations = []
@@ -265,66 +267,66 @@ class Test_globaldb_tests(unittest.TestCase):
         write_endpoint, read_endpoint = client._global_endpoint_manager.UpdateLocationsCache(writable_locations, readable_locations)
 
         # If writable_locations and readable_locations are empty, both Read and Write Endpoints point to endpoint passed while creating the client instance
-        self.assertEqual(write_endpoint, host)
-        self.assertEqual(read_endpoint, host)
+        self.assertEqual(write_endpoint, Test_globaldb_tests.host)
+        self.assertEqual(read_endpoint, Test_globaldb_tests.host)
 
-        writable_locations = [{'name' : write_location, 'databaseAccountEndpoint' : write_location_host}]
+        writable_locations = [{'name' : Test_globaldb_tests.write_location, 'databaseAccountEndpoint' : Test_globaldb_tests.write_location_host}]
         readable_locations = []
 
         write_endpoint, read_endpoint = client._global_endpoint_manager.UpdateLocationsCache(writable_locations, readable_locations)
 
         # If there are no readable_locations, we use the write endpoint as ReadEndpoint
-        self.assertEqual(write_endpoint, write_location_host)
-        self.assertEqual(read_endpoint, write_location_host)
+        self.assertEqual(write_endpoint, Test_globaldb_tests.write_location_host)
+        self.assertEqual(read_endpoint, Test_globaldb_tests.write_location_host)
 
         writable_locations = []
-        readable_locations = [{'name' : read_location, 'databaseAccountEndpoint' : read_location_host}]
+        readable_locations = [{'name' : Test_globaldb_tests.read_location, 'databaseAccountEndpoint' : Test_globaldb_tests.read_location_host}]
 
         write_endpoint, read_endpoint = client._global_endpoint_manager.UpdateLocationsCache(writable_locations, readable_locations)
 
         # If there are no writable_locations, both Read and Write Endpoints point to endpoint passed while creating the client instance
-        self.assertEqual(write_endpoint, host)
-        self.assertEqual(read_endpoint, host)
+        self.assertEqual(write_endpoint, Test_globaldb_tests.host)
+        self.assertEqual(read_endpoint, Test_globaldb_tests.host)
 
-        writable_locations = [{'name' : write_location, 'databaseAccountEndpoint' : write_location_host}]
-        readable_locations = [{'name' : read_location, 'databaseAccountEndpoint' : read_location_host}, {'name' : read_location2, 'databaseAccountEndpoint' : read_location2_host}]
+        writable_locations = [{'name' : Test_globaldb_tests.write_location, 'databaseAccountEndpoint' : Test_globaldb_tests.write_location_host}]
+        readable_locations = [{'name' : Test_globaldb_tests.read_location, 'databaseAccountEndpoint' : Test_globaldb_tests.read_location_host}, {'name' : Test_globaldb_tests.read_location2, 'databaseAccountEndpoint' : Test_globaldb_tests.read_location2_host}]
 
         connection_policy = documents.ConnectionPolicy()
-        connection_policy.PreferredLocations = [read_location2]
+        connection_policy.PreferredLocations = [Test_globaldb_tests.read_location2]
 
-        client = document_client.DocumentClient(host, {'masterKey': masterKey}, connection_policy)
+        client = document_client.DocumentClient(Test_globaldb_tests.host, {'masterKey': Test_globaldb_tests.masterKey}, connection_policy)
 
         write_endpoint, read_endpoint = client._global_endpoint_manager.UpdateLocationsCache(writable_locations, readable_locations)
 
         # Test that the preferred location is set as ReadEndpoint instead of default write endpoint when no preference is set
-        self.assertEqual(write_endpoint, write_location_host)
-        self.assertEqual(read_endpoint, read_location2_host)
+        self.assertEqual(write_endpoint, Test_globaldb_tests.write_location_host)
+        self.assertEqual(read_endpoint, Test_globaldb_tests.read_location2_host)
 
-        writable_locations = [{'name' : write_location, 'databaseAccountEndpoint' : write_location_host}, {'name' : read_location2, 'databaseAccountEndpoint' : read_location2_host}]
-        readable_locations = [{'name' : read_location, 'databaseAccountEndpoint' : read_location_host}]
+        writable_locations = [{'name' : Test_globaldb_tests.write_location, 'databaseAccountEndpoint' : Test_globaldb_tests.write_location_host}, {'name' : Test_globaldb_tests.read_location2, 'databaseAccountEndpoint' : Test_globaldb_tests.read_location2_host}]
+        readable_locations = [{'name' : Test_globaldb_tests.read_location, 'databaseAccountEndpoint' : Test_globaldb_tests.read_location_host}]
 
         connection_policy = documents.ConnectionPolicy()
-        connection_policy.PreferredLocations = [read_location2]
+        connection_policy.PreferredLocations = [Test_globaldb_tests.read_location2]
 
-        client = document_client.DocumentClient(host, {'masterKey': masterKey}, connection_policy)
+        client = document_client.DocumentClient(Test_globaldb_tests.host, {'masterKey': Test_globaldb_tests.masterKey}, connection_policy)
 
         write_endpoint, read_endpoint = client._global_endpoint_manager.UpdateLocationsCache(writable_locations, readable_locations)
 
         # Test that the preferred location is chosen from the WriteLocations if it's not present in the ReadLocations
-        self.assertEqual(write_endpoint, write_location_host)
-        self.assertEqual(read_endpoint, read_location2_host)
+        self.assertEqual(write_endpoint, Test_globaldb_tests.write_location_host)
+        self.assertEqual(read_endpoint, Test_globaldb_tests.read_location2_host)
 
-        writable_locations = [{'name' : write_location, 'databaseAccountEndpoint' : write_location_host}]
-        readable_locations = [{'name' : read_location, 'databaseAccountEndpoint' : read_location_host}, {'name' : read_location2, 'databaseAccountEndpoint' : read_location2_host}]
+        writable_locations = [{'name' : Test_globaldb_tests.write_location, 'databaseAccountEndpoint' : Test_globaldb_tests.write_location_host}]
+        readable_locations = [{'name' : Test_globaldb_tests.read_location, 'databaseAccountEndpoint' : Test_globaldb_tests.read_location_host}, {'name' : Test_globaldb_tests.read_location2, 'databaseAccountEndpoint' : Test_globaldb_tests.read_location2_host}]
 
         connection_policy.EnableEndpointDiscovery = False
-        client = document_client.DocumentClient(host, {'masterKey': masterKey}, connection_policy)
+        client = document_client.DocumentClient(Test_globaldb_tests.host, {'masterKey': Test_globaldb_tests.masterKey}, connection_policy)
 
         write_endpoint, read_endpoint = client._global_endpoint_manager.UpdateLocationsCache(writable_locations, readable_locations)
 
         # If EnableEndpointDiscovery is False, both Read and Write Endpoints point to endpoint passed while creating the client instance
-        self.assertEqual(write_endpoint, host)
-        self.assertEqual(read_endpoint, host)
+        self.assertEqual(write_endpoint, Test_globaldb_tests.host)
+        self.assertEqual(read_endpoint, Test_globaldb_tests.host)
 
     def test_globaldb_locational_endpoint_parser(self):
         url_endpoint='https://contoso.documents.azure.com:443/'
@@ -342,13 +344,19 @@ class Test_globaldb_tests(unittest.TestCase):
         self.assertEqual(locational_endpoint, 'https://contoso-EastUS.documents.azure.com:443/')
 
     def test_globaldb_endpoint_discovery_retry_policy_mock(self):
-        client = document_client.DocumentClient(host, {'masterKey': masterKey})
+        client = document_client.DocumentClient(Test_globaldb_tests.host, {'masterKey': Test_globaldb_tests.masterKey})
 
         self.OriginalExecuteFunction = endpoint_discovery_retry_policy._ExecuteFunction
         endpoint_discovery_retry_policy._ExecuteFunction = self._MockExecuteFunction
 
         self.OriginalGetDatabaseAccount = client.GetDatabaseAccount
         client.GetDatabaseAccount = self._MockGetDatabaseAccount
+
+        max_retry_attempt_count = 10
+        retry_after_in_milliseconds = 500
+
+        endpoint_discovery_retry_policy._EndpointDiscoveryRetryPolicy.Max_retry_attempt_count = max_retry_attempt_count
+        endpoint_discovery_retry_policy._EndpointDiscoveryRetryPolicy.Retry_after_in_milliseconds = retry_after_in_milliseconds
 
         document_definition = { 'id': 'doc',
                                 'name': 'sample document',